requirements.txt
customtkinter==5.2.1
pillow>=10.3.0
pandas==2.1.3
openpyxl==3.1.2
pytesseract==0.3.10
python-dateutil==2.8.2
